refactor file names as kebab-case for consistency
