+++ src/layouts/index.ts
@@ -1,4 +1,4 @@
-export { default as App } from "./AppLayout"
-export { default as Auth } from "./AuthLayout"
-export { default as Root } from "./RootLayout"
-export { default as Tenant } from "./TenantLayout"
+export { default as App } from "./app-layout"
+export { default as Auth } from "./auth-layout"
+export { default as Root } from "./root-layout"
+export { default as Tenant } from "./tenant-layout"
