+++ src/pages/app/index.ts
@@ -1,1 +1,1 @@
-export { default as Dashboard } from "./Dashboard"
+export { default as Dashboard } from "./dashboard"
