+++ src/pages/auth/Login.tsx
@@ -16,8 +16,8 @@ import {
 import { Input } from "@/components/ui/input"
 
 import * as keygen from "@/keygen/index"
-import { useAuth } from "@/hooks/useAuth"
-import * as Loading from "@/components/Loading"
+import { useAuth } from "@/hooks/use-auth"
+import * as Loading from "@/components/loading"
 
 const emailSchema = z.object({
   username: z.string().email("Please enter a valid email."),
