+++ src/pages/auth/index.ts
@@ -1,8 +1,8 @@
-export { default as Login } from "./Login"
-export { default as Password } from "./Password"
-export { default as SSO } from "./SSO"
-export { default as Verify } from "./Verify"
-export { default as Register } from "./Register"
-export { default as Recovery } from "./Recovery"
-export { default as Sent } from "./Sent"
-export { default as Reset } from "./Reset"
+export { default as Login } from "./login"
+export { default as Password } from "./password"
+export { default as SSO } from "./sso"
+export { default as Verify } from "./verify"
+export { default as Register } from "./register"
+export { default as Recovery } from "./recovery"
+export { default as Sent } from "./sent"
+export { default as Reset } from "./reset"
