+++ src/pages/index.ts
@@ -4,5 +4,5 @@ import * as Auth from "./auth/index"
 export { App }
 export { Auth }
 
-export { default as Redirect } from "./Redirect"
-export { default as Error } from "./Error"
+export { default as Redirect } from "./redirect"
+export { default as Error } from "./error"
